Fix: NumPy data type error in audio pause generation
🔧 **Problem Fixed:**
- Error: Cannot interpret 'torch.float32' as a data type
- Occurred during audio chunk concatenation with pauses

✅ **Solution:**
- Changed pause generation from `dtype=wav.dtype` to `dtype=np.float32`
- Fixed both PDF and Direct TTS functions
- Ensures consistent NumPy data type specification

🎯 **Result:**
- Audio generation now works without data type errors
- Long text chunking functions properly
- Natural pauses between chunks work correctly

+++ main.py
@@ -290,7 +290,7 @@ def generate_audio_from_text(
             
             # Add a small pause between chunks (0.3 seconds of silence)
             pause_samples = int(0.3 * model.sr)
-            pause = np.zeros((1, pause_samples), dtype=wav.dtype)
+            pause = np.zeros((1, pause_samples), dtype=np.float32)
             
             audio_segments.append(wav.squeeze(0).numpy())
             if i < len(text_chunks) - 1:  # Don't add pause after last chunk
@@ -348,7 +348,7 @@ def generate_from_text(
                 
                 # Add small pause between chunks
                 pause_samples = int(0.3 * model.sr)
-                pause = np.zeros((1, pause_samples), dtype=wav.dtype)
+                pause = np.zeros((1, pause_samples), dtype=np.float32)
                 
                 audio_segments.append(wav.squeeze(0).numpy())
                 if i < len(text_chunks) - 1:
